Adding list of algorithms.
