+++ share/eows/config/algorithms.json
@@ -1,0 +1,23 @@
+{
+	"algorithm": "TWDTW",
+	"input_parameters": {
+		"coverage": "mod13q1",
+		"attributes": ["ndvi"],
+		"roi": {
+			"geometry": {
+				"type": "Polygon",
+				"coordinates": [
+					[[-10.0, -10.0], [10.0, -10.0], [10.0, 10.0], [-10.0, -10.0]]
+				]
+			}
+		},
+		"patterns": "patterns",
+		"start_date": "2000-09-01",
+		"end_date": "2016-08-31",
+		"by": "12 month",
+		"overlap": 0.5,
+		"alpha": -0.1,
+		"beta": 100.0,
+		"output_array_name": "mod13q1_luc"
+ }
+}
