+++ src/eows/wtscs/wtscs.cpp
@@ -17,15 +17,6 @@
   e-sensing team at <esensing-team@dpi.inpe.br>.
  */
 
-/*!
-  \file eows/wtscs/wtscs.cpp
-
-  \brief Web Time Series Classification Service.
-
-  \author Eduardo Llapa Rodriguez
-  \author Gilberto Ribeiro de Queiroz
- */
-
 // EOWS
 #include "wtscs.hpp"
 #include "../core/http_response.hpp"
@@ -34,10 +25,10 @@
 #include "../core/service_operations_manager.hpp"
 #include "../core/utils.hpp"
 #include "../geoarray/data_types.hpp"
-#include "../exception.hpp"
+//#include "../exception.hpp"
 
-// C++ Standard Library
-#include <memory>
+//// C++ Standard Library
+//#include <memory>
 
 // Boost
 #include <boost/format.hpp>
@@ -47,43 +38,23 @@
 #include <rapidjson/stringbuffer.h>
 #include <rapidjson/writer.h>
 
-static void
-return_exception(const char* msg, eows::core::http_response& res);
+static void return_exception(const char* msg, eows::core::http_response& res);
 
-namespace eows
+void eows::wtscs::status_handler::do_get(const eows::core::http_request& req, eows::core::http_response& res)
 {
-  namespace wtscs
-  {
-    struct base_input_parameters
-    {
-      virtual ~base_input_parameters() = default;
-    };
+//  TODO: See how to implement the status_handler methods!.
 
-    struct twdtw_input_parameters : base_input_parameters
-    {
-      std::string cv_name;
-      std::vector<std::string> attributes;
-      eows::geoarray::spatial_extent_t roi;
-      std::string start_date;
-      std::string end_date;
-      std::string by;
-      double overlap;
-      double alpha;
-      double beta;
-      std::string output_array_name;
-    };
+  std::string return_msg("\"status_handler\"");
 
-    struct classify_request_parameters
-    {
-      std::string algorithm;
-      std::unique_ptr<base_input_parameters> input_parameters;
-    };
+  res.set_status(eows::core::http_response::OK);
 
-  }
+  res.add_header(eows::core::http_response::CONTENT_TYPE, "text/plain; charset=utf-8");
+  res.add_header(eows::core::http_response::ACCESS_CONTROL_ALLOW_ORIGIN, "*");
+
+  res.write(return_msg.c_str(), return_msg.size());
 }
 
-void eows::wtscs::list_algorithms_handler::do_get(const eows::core::http_request& req,
-                                                  eows::core::http_response& res)
+void eows::wtscs::list_algorithms_handler::do_get(const eows::core::http_request& req, eows::core::http_response& res)
 {
     std::vector<std::string> algorithms{ "twdtw", "bfast", "bfast-monitor" };
 
@@ -133,9 +104,31 @@ void eows::wtscs::classify_handler::do_post(const eows::core::http_request& req,
     if((jalgorithm == doc.MemberEnd()) || (!jalgorithm->value.IsString()))
       throw eows::parse_error("Please check key 'algorithm' in WTSCS classify request.");
 
-    req_parameters.algorithm = jalgorithm->value.GetString();
+//  TODO: To get and check up the input parameters.
 
-// assembly the response
+//  If algorithm exist then extract the specific parameters.
+//
+//  req_parameters.algorithm = jalgorithm->value.GetString();
+
+//  rapidjson::Value::ConstMemberIterator jCoverage = doc.FindMember("coverage");
+//  if((jInputParameters == doc.MemberEnd()) || (!jInputParameters->value.IsString()))
+//  throw eows::parse_error("Please check key 'coverage' in WTSCS classify request.");
+//  ...
+//  Other case:
+//  eows::parse_error("Algortihm no found.");
+
+//  Obs: Store the patterns file in a directory visible by SciDB.
+
+
+
+//  TODO: Generate the  UUID identifier to create the AFL syntax.
+
+//  It Sends the AFL request.
+//  The answer is a URL wrapping the request UUID identifier.
+//  Ex. http://localhost:7654/wtscs/status?UUID=123456687
+
+
+//  assembly the response
     rapidjson::StringBuffer buff;
     rapidjson::Writer<rapidjson::StringBuffer> writer(buff);
 
@@ -163,26 +156,23 @@ void eows::wtscs::classify_handler::do_post(const eows::core::http_request& req,
   }
 }
 
-void
-eows::wtscs::initialize()
+void eows::wtscs::initialize()
 {
   EOWS_LOG_INFO("Initializing WTSCS...");
 
-  {
-    std::unique_ptr<list_algorithms_handler> h(new list_algorithms_handler);
-    eows::core::service_operations_manager::instance().insert("/wtscs/list_algorithms", std::move(h));
-  }
+  std::unique_ptr<eows::wtscs::status_handler> s_h(new eows::wtscs::status_handler);
+  eows::core::service_operations_manager::instance().insert("/wtscs/status", std::move(s_h));
 
-  {
-    std::unique_ptr<classify_handler> h(new classify_handler);
-    eows::core::service_operations_manager::instance().insert("/wtscs/classify", std::move(h));
-  }
+  std::unique_ptr<list_algorithms_handler> la_h(new list_algorithms_handler);
+  eows::core::service_operations_manager::instance().insert("/wtscs/list_algorithms", std::move(la_h));
+
+  std::unique_ptr<classify_handler> c_h(new classify_handler);
+  eows::core::service_operations_manager::instance().insert("/wtscs/classify", std::move(c_h));
 
   EOWS_LOG_INFO("WTSCS service initialized!");
 }
 
-void
-return_exception(const char* msg, eows::core::http_response& res)
+void return_exception(const char* msg, eows::core::http_response& res)
 {
   rapidjson::StringBuffer buff;
   rapidjson::Writer<rapidjson::StringBuffer> writer(buff);
